refactor: consolidate resolvers into domain-focused modules
Split resolvers.ts into theme.ts and language.ts for better organization:
- theme.ts: resolveTheme, toShikiOptions, getMultiThemeOptions, DEFAULT_THEMES
- language.ts: resolveLanguage, isPlaintext helper
- options.ts: simplified to import from theme.ts
