+++ package/src/lib/options.ts
@@ -1,40 +1,13 @@
-import type {
-  CodeToHastOptions,
-  CodeOptionsSingleTheme,
-  CodeOptionsMultipleThemes,
-  BundledTheme,
-} from 'shiki';
+import type { CodeToHastOptions } from 'shiki';
 
-import type { HighlighterOptions, OutputFormat, Themes } from './types';
-import type { ThemeResult } from './resolvers';
+import type { HighlighterOptions, OutputFormat } from './types';
+import type { ThemeResolution } from './theme';
+import { toShikiOptions, getMultiThemeOptions } from './theme';
 import { lineNumbersTransformer } from './transformers';
-
-const DEFAULT_THEMES: Themes = {
-  light: 'github-light',
-  dark: 'github-dark',
-};
-
-const buildThemeOptions = (
-  themeResult: ThemeResult
-):
-  | CodeOptionsSingleTheme<BundledTheme>
-  | CodeOptionsMultipleThemes<BundledTheme> => {
-  const { isMultiTheme, multiTheme, singleTheme } = themeResult;
-
-  if (isMultiTheme) {
-    return {
-      themes: multiTheme || DEFAULT_THEMES,
-    } as CodeOptionsMultipleThemes<BundledTheme>;
-  }
-
-  return {
-    theme: singleTheme || DEFAULT_THEMES.dark,
-  } as CodeOptionsSingleTheme<BundledTheme>;
-};
 
 export const buildShikiOptions = <F extends OutputFormat>(
   languageId: string,
-  themeResult: ThemeResult,
+  themeResolution: ThemeResolution,
   options: HighlighterOptions<F>
 ): CodeToHastOptions => {
   const {
@@ -51,11 +24,6 @@ export const buildShikiOptions = <F extends OutputFormat>(
     ...restOptions
   } = options;
 
-  const themeOptions = buildThemeOptions(themeResult);
-  const multiThemeOptions = themeResult.isMultiTheme
-    ? { defaultColor, cssVariablePrefix }
-    : {};
-
   const transformers = [...(restOptions.transformers || [])];
   if (showLineNumbers && outputFormat !== 'tokens') {
     transformers.push(lineNumbersTransformer(startingLineNumber));
@@ -63,11 +31,13 @@ export const buildShikiOptions = <F extends OutputFormat>(
 
   return {
     lang: languageId,
-    ...themeOptions,
-    ...multiThemeOptions,
+    ...toShikiOptions(themeResolution),
+    ...getMultiThemeOptions(
+      themeResolution,
+      defaultColor,
+      cssVariablePrefix
+    ),
     ...restOptions,
     transformers,
   };
 };
-
-export { DEFAULT_THEMES };
